+++ src/components/index.js
@@ -4,3 +4,4 @@ export { default as ListOfNotes } from './ListOfNotes/ListOfNotes.js';
 export { default as TreeView } from './TreeView/TreeView.js';
 export { default as ContextMenu } from './TreeView/ContextMenu.js';
 export { default as useRtlDirection } from './Redactor/useRtlDirection.js';
+export { calculateRtlDirection } from './helpers.js';
